fix: sprites being offset by 1px when their DrawPos is negative
Raylib truncates all values before drawing. This is uneventful for us,
because wrapping from the right and bottom of the world causes the
sprite to have negative coordinates. This problem only arised with
non-integral, negative float values. We are now explicitly flooring the
coordinates before drawing to avoid this awkward edge case.